Table 1: JSON Web Token (JWT) restricted claim set
Request an update to the mentioned page to somehow include the "name" claim as a restricted claim that cannot be excluded. Because currently there is no reference to such claim and there are references to others related as basic claims, it creates the sense that it can be manipulated/excluded.