Payload claims
Where it says that "The name claim provides a human-readable value that identifies the subject of the token. The value isn't guaranteed to be unique, it can be changed, and it's designed to be used only for display purposes. The profile scope is required to receive this claim."
As already presented above, this is not accurate and "name" claim will always be present.